Check whether generated inputs already belong to the test set (disabled by default for now), and dump test origins
The membership test relies on hashing of numpy arrays, so actually new
tests might actually be discarded in case of hash collision (i.e,
regarded as already in the test set).  It is still not clear how
frequent/likely such collisions are.

Test origins are dumped as a YAML file by the engine at the end of
test case generation: this is essentially an array that maps each test
case index with the index of the test case it originates from.